An attempt to extend country.subdivision
This approach feels a tad awkward, and I may revert many of these
changes. But I wanted to try this out and commit to the repository, at
least for posterity. Perhaps it is just different and will take awhile
to get used to.

Though this approach has the benefits of not creating an extra
model/table and takes advantage of other modules (hypothetically) that
use or extend `country.subdivision` for one reason or another (see
`account_stock_eu` for one extension), here are some potential
downsides:

- `Subdivision.rec_name` with the FIPS code feels out-of-place for
  addresses (a common use case for Subdivision right now).
- We are dealing with a different dataset and standard than the ISO 3166
  standard embodied by the pycountry module, and some data are
  represented differently conceptually.
- Creating a new model affords a simpler design, especially when
  relating to US Census regions and organizations of subdivisions.
- Generalizing a class code into a `Subdivision.type` was interesting,
  but may be too lossy for some users.